refactor: fix inline usages
