+++ src/main/kotlin/nz/adjmunro/knomadic/outcome/Maybe.kt
@@ -7,10 +7,10 @@ import nz.adjmunro.knomadic.outcome.members.mapFailure
 import nz.adjmunro.knomadic.raise.RaiseScope
 import nz.adjmunro.knomadic.raise.RaiseScope.Companion.default
 import nz.adjmunro.knomadic.raise.RaiseScope.Companion.fold
-import nz.adjmunro.knomadic.inline.nullfold
-import nz.adjmunro.knomadic.inline.rethrow
-import nz.adjmunro.knomadic.inline.throwfold
-import nz.adjmunro.knomadic.inline.unit
+import nz.adjmunro.inline.nullfold
+import nz.adjmunro.inline.rethrow
+import nz.adjmunro.inline.throwfold
+import nz.adjmunro.inline.unit
 
 @KnomadicDsl
 public typealias Maybe<Ok> = Outcome<Ok, Unit>
