+++ src/main/kotlin/nz/adjmunro/knomadic/outcome/members/Fold.kt
@@ -2,7 +2,7 @@ package nz.adjmunro.knomadic.outcome.members
 
 import nz.adjmunro.knomadic.KnomadicDsl
 import nz.adjmunro.knomadic.outcome.Outcome
-import nz.adjmunro.knomadic.inline.itself
+import nz.adjmunro.inline.itself
 import kotlin.contracts.InvocationKind.AT_MOST_ONCE
 import kotlin.contracts.contract
